teosar/periodogram: Add alternative OpenCL periodogram code
The code enables to create your own periodogram functions,
and test in parallel on GPU or CPU for a lot of PS many
input values in order to maximize the periodogram.

This implementation currently only supports the simple
periodogram with 2 variables to optimize, but the core
of the code supports more, and it should be simple to
add support for more configurations.