teosar:workflow: Catch Secondary Exceptions
We expect the secondary pipeline to run on many products, thus, instead of interrupting the program if a problem occurs for a product, we catch the exception  in the execute function and issue a Warning. This choice is debatable and lies on the assumption that a few products might raise exceptions, but we might want the result on the other products anyway.